requirements.txt
torch==1.8.0
torchmetrics==0.2.0
torchsummary==1.5.1
pytorch-lightning==1.2.10
torchvision==0.9.0
tqdm==4.46.1
Pillow==7.1.2
easing-functions==1.0.4
opencv-python==4.2.0.34
numpy==1.19.4
scikit-image==0.17.2
scikit-learn==0.24.2
ipython==7.15.0
tensorboard==2.4.1
lpips==0.1.3